requirements.txt
zhdate==0.1
rsa==4.7.2
requests==2.25.1
pipreqs==0.4.10
PyQRCode==1.2.1
sty==1.0.0rc1
fpdf==1.7.2
pycryptodome==3.10.1
pyDes==2.0.1
PyJWT==2.1.0
PyPDF2==1.26.0
pypng==0.0.20
